+++ src/common/utxobased/engine/ServerStates.ts
@@ -409,7 +409,7 @@ export function makeServerStates(config: ServerStateConfig): ServerStates {
           //   /^http(?:s)?:/i
           // ])
 
-          const { nowNodeApiKey } = initOptions
+          const { nowNodesApiKey } = initOptions
           const nowNodeUris = serverConfigs
             .filter(config => config.type === 'blockbook-nownode')
             .map(config => config.uris)
@@ -429,7 +429,7 @@ export function makeServerStates(config: ServerStateConfig): ServerStates {
           }
 
           // If there is no key for the NowNode servers:
-          if (nowNodeApiKey == null) {
+          if (nowNodesApiKey == null) {
             reject(new Error('Missing connection key for fallback servers.'))
             return
           }
@@ -441,7 +441,7 @@ export function makeServerStates(config: ServerStateConfig): ServerStates {
             io.fetchCors(`${uri}/api/v2/sendtx/`, {
               method: 'POST',
               headers: {
-                'api-key': nowNodeApiKey
+                'api-key': nowNodesApiKey
               },
               body: transaction.signedTx
             })